Fix #9, handle API version changes on image refresh
If the image parser encounters a file that doesn't have the same "api_version" it will try to refresh the default ones. If it still cannot parse the file, it will throw an error (exit 1).
Update default data path to "wilfred-data" insteaf of "wilfred" because some shells will try to cd into the folder instead of running the command.